feat: route only x86/x64 PE/ELF to the BCJ bucket
bucket_for_bytes sent every PE/ELF to bucket 1, which is finalized with
xz --x86 -- an ARM64 ELF or ARM64 PE got the x86 branch filter applied
(reversible, but costs ratio on foreign branch encodings). Now parses
the PE machine field (e_lfanew+4; keep 0x014C/0x8664) and ELF e_machine
(offset 18, endian-aware; keep EM_386/EM_X86_64); everything else goes
to bucket 0. Bare-DOS MZ stays bucket 1 (x86 by construction).

Synthetic check: fake ARM64 ELF/PE now pack single-bucket, fake x64 ELF
still routes to bucket 1; all RT OK. x86 bench fixtures byte-identical
(pe-deflate.zip 1,755,573 = master). Also fixes the %zu format warning
in verify_entry (MinGW printf lacks %z).

+++ src/zxle.c
@@ -592,8 +592,8 @@ static void verify_entry(const char *path, uint32_t want_crc, uint64_t want_size
     uint32_t c = crc32_buf(got, got_n);
     free(got);
     if ((uint64_t)got_n != want_size) {
-        fprintf(stderr, "zxle: %s: size %zu != %llu\n", path, got_n,
-                (unsigned long long)want_size);
+        fprintf(stderr, "zxle: %s: size %llu != %llu\n", path,
+                (unsigned long long)got_n, (unsigned long long)want_size);
         die("entry size mismatch");
     }
     if (c != want_crc) {
